requirements.txt
pytest~=6.2.5
pytest-html
selenium~=3.141.0
Appium-Python-Client~=1.2.0
py~=1.10.0
xlrd~=2.0.1
ruamel.yaml